Migrating the UI Feature components to new locations in the Loopback-inspired project structure
Ths time, no bad path mishaps.

Only one file was not recognized as having a history--angular-csp.css.
It's an obligatory CSS file that is pacakged (normally) with angularJS.
THis seems not too bad, as I want to start getting this file from the
/vendor distribution anyhow.

More information to come as the work progresses into an example, but for
now, there are still framework components to migrate.