Getting development index.html from staging area
This is a step towards using a concattenated but not minified script
file in development.

Initially done with a view that I'd need browserification for non-LB
code as well, but may not be so important anymore since I'm going to
revise all such dependencies to use a global window element rather than
relying on nodeJS require() semantics.

Will probably revert this soon enough.