Restoring more features
Progress towards getting livereload functional, as well as build-time
injection of a list of our own JS to augment the vendor set generated
andadded by grunt-wiredep.  Uses grunt-html-build (NOT grunt-
htmlbuild!!!)

Adding nodemon/node-inspector and some watch logic to restart the server
on appropraite events.  Again, something of a work in progress, but in a
st ate worth checkpointing!