Improved Live Relooad functionality
-- realized the reason LiveReload failed to configure it client side
websocket fiarly often was because the script used to plant a tag that
downloads the LiveReload client library relies on finding its own tag in
the HTML to initialize, but could not do so becaus eI'd named the script
that plans that tag too similarly and so it was finding the wrong tag.

Corrected htat by renaming livereload.js to lr_init.js.  Also removed
the duplicate copy that was incorectly getting addred to the BRowserify
bundle.  Reversed a number of patchy workaround that had not been the
prooper solutions one these rioot causes were correctyed.