Got coffeify and remapify cooperating at last!
An attempt to resolve the issue through Gruntfile was not successful.
Giving Coffeeify a configuration object, even a blank one, much less one
with { global: true }, was a gross dead end that mucked up Browserify's
internal "model-deps" subsystem.

What did work was adding a the following content to package.json:

+  "browser": "./client/ngapp/scripts/jchptf/jchptf.js",
+  "browserify": {
+    "transform": [
+      "coffeeify"
+    ]
+  },

...after which it apparently becomes necessary to register coffeeify as
a runtime dependency in package.json and not just a dev dependency.
There is possibly a way to fix the build-only use route, especially
since that works fine with remapify removed, but remapify turns out to
be too valuable for applying a naming convention to the require
artifacts, esppecially important for allowing the client code base to
access anything from the common code base!

    Until that solution is available, this workaround is a fine thing.