Prefixing server side boot scripts to set a fixed execution order.
Loopback executes its boot script dir's content according to
lexographical order, so there is a best practice recommendation to mimic
Linux rc.d numeric name prefixing to control boot order.  I should do
this a second time later on to create gaps in the sequence for future
ordering, but I skipped past that this time and just focused on getting
the ordering useful.