Crossword editor view model semantics
Fleshed out transient derived client models that will hold the UI state
of the running crossword editor use cases.  The persistence model is
compact, but the UI needs a model that has discrete packets of state for
each distinct UI element, namely the cursor location, word highlight
boundaries, individual cell contents.  It also needs to be able to
analyze the current state for errors and detection of grid content that
is ready to be used for the second editting mode, tasked with enhancing
the prize revelation experience.

Also removed a vestigial "user.json" model that was just a misnamed
collection of properties not intended for real use.