Build system fixes
Corrected browserify use to correct a confusion as to the difference
between add() and require().  Had to swap their respective uses to
correct.

Fixed the jade pipeline to account for additional postprocessing that only
affects index.html (or index.jade)

Corrected some of the file signalling between watch and nodemon.

Added a concatentaiton of appConfig.source.client to appConfig.app and
called it appConfig.source.app.  Simplification for a common use case.

Added temporary debug statements to the "bundleClient" task as a precursor
for making the selective source map extraction work correctly during a
build.